- Changing INT columns to BIGINT.

+++ data/schema/schema.mysql.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `user` (
-    `user_id`      INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `user_id`      BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `username`     VARCHAR(255) DEFAULT NULL UNIQUE,
     `email`        VARCHAR(255) DEFAULT NULL UNIQUE,
     `display_name` VARCHAR(50) DEFAULT NULL,
@@ -8,7 +8,7 @@ CREATE TABLE `user` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE `user_role_linker` (
-    `user_id` INT(11) UNSIGNED NOT NULL,
+    `user_id` BIGINT(20) UNSIGNED NOT NULL,
     `role_id` VARCHAR(45) NOT NULL,
     PRIMARY KEY (`user_id`,`role_id`),
     INDEX `user_role_linker_fk1` (`user_id`),
@@ -16,20 +16,20 @@ CREATE TABLE `user_role_linker` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE `team` (
-    `id`   INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id`   BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `name` VARCHAR(255) DEFAULT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE `user_settings` (
-    `user_id` INT(11) UNSIGNED NOT NULL PRIMARY KEY,
+    `user_id` BIGINT(20) UNSIGNED NOT NULL PRIMARY KEY,
     `locale`  VARCHAR(20) NOT NULL, -- Currently known max length is 11 char.
-    `team_id` INT(11) UNSIGNED DEFAULT NULL,
+    `team_id` BIGINT(20) UNSIGNED DEFAULT NULL,
     CONSTRAINT `user_settings_fk1` FOREIGN KEY (`user_id`) REFERENCES `user` (`user_id`) ON DELETE CASCADE ON UPDATE CASCADE,
     CONSTRAINT `user_settings_fk2` FOREIGN KEY (`team_id`) REFERENCES `team` (`id`) ON DELETE SET NULL ON UPDATE CASCADE
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE `user_languages` (
-    `user_id` INT(11) UNSIGNED NOT NULL,
+    `user_id` BIGINT(20) UNSIGNED NOT NULL,
     `locale`  VARCHAR(20) NOT NULL, -- Currently known max length for primary locale is 3 char.
     PRIMARY KEY (`user_id`,`locale`),
     INDEX `user_languages_fk1` (`user_id`),
@@ -37,8 +37,8 @@ CREATE TABLE `user_languages` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE `team_member` (
-    `user_id` INT(11) UNSIGNED NOT NULL,
-    `team_id` INT(11) UNSIGNED NOT NULL,
+    `user_id` BIGINT(20) UNSIGNED NOT NULL,
+    `team_id` BIGINT(20) UNSIGNED NOT NULL,
     PRIMARY KEY (`user_id`,`team_id`),
     INDEX `team_member_fk1` (`user_id`),
     INDEX `team_member_fk2` (`team_id`),
@@ -47,8 +47,8 @@ CREATE TABLE `team_member` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE `app` (
-    `id`                 INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    `team_id`            INT(11) UNSIGNED DEFAULT NULL,
+    `id`                 BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `team_id`            BIGINT(20) UNSIGNED DEFAULT NULL,
     `name`               VARCHAR(255) DEFAULT NULL,
     `path_to_res_folder` VARCHAR(4096) DEFAULT NULL,
     `git_repository`     VARCHAR(4096) DEFAULT NULL,
@@ -61,8 +61,8 @@ CREATE TABLE `app` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE `app_resource` (
-    `id`             INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    `app_id`         INT(11) UNSIGNED NOT NULL,
+    `id`             BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `app_id`         BIGINT(20) UNSIGNED NOT NULL,
     `name`           VARCHAR(255) NOT NULL,
     `locale`         VARCHAR(20) NOT NULL,
     `primary_locale` VARCHAR(20) NOT NULL, -- Currently known max length for primary locale is 3 char. Field isn't available in model.
@@ -74,8 +74,8 @@ CREATE TABLE `app_resource` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE `app_resource_file` (
-    `id`     INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    `app_id` INT(11) UNSIGNED NOT NULL,
+    `id`     BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `app_id` BIGINT(20) UNSIGNED NOT NULL,
     `name`   VARCHAR(255) NOT NULL,
     INDEX `app_resource_file_fk1` (`app_id`),
     UNIQUE INDEX `app_resource_file_uk1` (`app_id`, `name`),
@@ -83,7 +83,7 @@ CREATE TABLE `app_resource_file` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE `resource_type` (
-    `id`        INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id`        BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `name`      VARCHAR(255) NOT NULL,
     `node_name` VARCHAR(255) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
@@ -91,9 +91,9 @@ CREATE TABLE `resource_type` (
 INSERT INTO `resource_type` (`id`, `name`, `node_name`) VALUES (1, 'String', 'string');
 
 CREATE TABLE `resource_file_entry` (
-    `id`                   INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    `app_resource_file_id` INT(11) UNSIGNED DEFAULT NULL,
-    `resource_type_id`     INT(11) UNSIGNED NOT NULL,
+    `id`                   BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `app_resource_file_id` BIGINT(20) UNSIGNED DEFAULT NULL,
+    `resource_type_id`     BIGINT(20) UNSIGNED NOT NULL,
     `name`                 VARCHAR(255) NOT NULL,
     `deleted`              TINYINT(1) NOT NULL,
     `translatable`         TINYINT(1) NOT NULL,
@@ -116,9 +116,9 @@ $$
 DELIMITER ;
 
 CREATE TABLE `resource_file_entry_string` (
-    `id`                     INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    `app_resource_id`        INT(11) UNSIGNED NOT NULL,
-    `resource_file_entry_id` INT(11) UNSIGNED NOT NULL,
+    `id`                     BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `app_resource_id`        BIGINT(20) UNSIGNED NOT NULL,
+    `resource_file_entry_id` BIGINT(20) UNSIGNED NOT NULL,
     `value`                  VARCHAR(20480) NOT NULL,
     `last_change`            BIGINT(20) NOT NULL,
     INDEX `resource_file_entry_string_ik1` (`last_change`),
